Move method configuration widgets to this repo
