--- nbmolviz/markdown/__init__.py
@@ -1,1 +1,0 @@
-from .notebook_display import *
