+++ nbmolviz/mixins/__init__.py
@@ -1,0 +1,4 @@
+from .mixins import *
+from .atoms import *
+from .molecule import *
+from .method import *
