+++ nbmolviz/mixins/atoms.py
@@ -18,9 +18,7 @@ import IPython.display
 from moldesign import data, utils
 from moldesign import units as u
 
-from .. import widgets as mdtwidgets
 from .. import viewer, uibase
-
 
 class AtomNotebookMixin(object):
     """ Notebook-specific methods for atoms.
@@ -216,221 +214,3 @@ class AtomGroupNotebookMixin(object):
         self.viz2d = viewer.DistanceGraphViewer(alist, **kwargs)
         if highlight_atoms: self.viz2d.highlight_atoms(highlight_atoms)
         return self.viz2d
-
-
-class ResidueNotebookMixin(object):
-    """ Mixin class for Residues with notebook-relevenat classes
-    """
-
-    def _repr_markdown_(self):
-        return self.markdown_summary()
-
-    def markdown_summary(self):
-        """ Markdown-formatted information about this residue
-
-        Returns:
-            str: markdown-formatted string
-        """
-        if self.type == 'placeholder':
-            return '`%s`' % repr(self)
-
-        if self.molecule is None:
-            lines = ["<h3>Residue %s</h3>" % self.name]
-        else:
-            lines = ["<h3>Residue %s (index %d)</h3>" % (self.name, self.index)]
-
-        if self.type == 'protein':
-            lines.append('**Residue codes**: %s / %s' % (self.resname, self.code))
-        else:
-            lines.append("**Residue code**: %s" % self.resname)
-        lines.append('**Type**: %s' % self.type)
-        if self.resname in data.RESIDUE_DESCRIPTIONS:
-            lines.append('**Description**: %s' % data.RESIDUE_DESCRIPTIONS[self.resname])
-
-        lines.append('**<p>Chain:** %s' % self.chain.name)
-
-        lines.append('**PDB sequence #**: %d' % self.pdbindex)
-
-        terminus = None
-        if self.type == 'dna':
-            if self.is_3prime_end:
-                terminus = "3' end"
-            elif self.is_5prime_end:
-                terminus = "5' end"
-        elif self.type == 'protein':
-            if self.is_n_terminal:
-                terminus = 'N-terminus'
-            elif self.is_c_terminal:
-                terminus = 'C-terminus'
-        if terminus is not None:
-            lines.append('**Terminal residue**: %s of chain %s' % (terminus, self.chain.name))
-
-        if self.molecule is not None:
-            lines.append("**Molecule**: %s" % self.molecule.name)
-
-        lines.append("**<p>Number of atoms**: %s" % self.num_atoms)
-        if self.backbone:
-            lines.append("**Backbone atoms:** %s" % ', '.join(x.name for x in self.backbone))
-            lines.append("**Sidechain atoms:** %s" % ', '.join(x.name for x in self.sidechain))
-        else:
-            lines.append("**Atom:** %s" % ', '.join(x.name for x in self.atoms))
-
-        return '<br>'.join(lines)
-
-
-class MolNotebookMixin(object):
-    """ Methods for displaying molecular information in Jupyter notebooks
-
-    Note:
-        This is a mixin class designed only to be mixed into the :class:`Molecule` class. Routines
-        are separated are here for code organization only - they could be included in the main
-        Molecule class without changing any functionality
-    """
-    def draw_orbitals(self, **kwargs):
-        """ Visualize any calculated molecular orbitals (Jupyter only).
-
-        Returns:
-            mdt.orbitals.OrbitalViewer
-        """
-        if 'wfn' not in self.properties:
-            self.calculate_wfn()
-        return mdtwidgets.orbitals.OrbitalViewer(self, **kwargs)
-
-
-    def markdown_summary(self):
-        """A markdown description of this molecule.
-
-        Returns:
-            str: Markdown"""
-        # TODO: remove leading underscores for descriptor-protected attributes
-        lines = ['### Molecule: "%s" (%d atoms)' % (self.name, self.natoms)]
-
-        description = self.metadata.get('description', None)
-        if description is not None:
-            description = self.metadata.description[:5000]
-            url = self.metadata.get('url', None)
-            if url is not None:
-                description = '<a href="%s" target="_blank">%s</a>' % \
-                              (url, description)
-            lines.append(description)
-
-        lines.extend([
-                 '**Mass**: {:.2f}'.format(self.mass),
-                 '**Formula**: %s' % self.get_stoichiometry(html=True),
-                 '**Charge**: %s' % self.charge])
-
-        if self.energy_model:
-            lines.append('**Potential model**: %s' % str(self.energy_model))
-
-        if self.integrator:
-            lines.append('**Integrator**: %s' % str(self.integrator))
-
-        if self.is_biomolecule:
-            lines.extend(self.biomol_summary_markdown())
-
-        return '\n\n'.join(lines)
-
-    def _repr_markdown_(self):
-        return self.markdown_summary()
-
-    def biomol_summary_markdown(self):
-        """A markdown description of biomolecular structure.
-
-        Returns:
-            str: Markdown string"""
-        lines = []
-        if len(self.residues) > 1:
-            table = self.get_residue_table()
-            lines.append('### Residues')
-            # extra '|' here may be workaround for a bug in ipy.markdown?
-            lines.append(table.markdown(replace={0: ' '}) + '|')
-
-            lines.append('### Biopolymer chains')
-            seqs = []
-            for chain in self.chains:
-                seq = chain._get_sequence(_html=True)
-                if not seq.strip():  # don't write anything if there's no sequence
-                    continue
-
-                seqs.append('**%s**: %s' % (chain.name, seq))
-            lines.append('<br>'.join(seqs))
-
-        return lines
-
-    def get_residue_table(self):
-        """Creates a data table summarizing this molecule's primary structure.
-
-        Returns:
-            moldesign.utils.MarkdownTable"""
-        table = utils.MarkdownTable(*(['chain'] +
-                                      'protein dna rna unknown water solvent'.split()))
-        for chain in self.chains:
-            counts = {}
-            unk = []
-            for residue in chain.residues:
-                cat = residue.type
-                if cat == 'unknown':
-                    unk.append(residue.name)
-                counts[cat] = counts.get(cat, 0) + 1
-            counts['chain'] = '<b>%s</b>' % chain.name
-            if 0 < len(unk) <= 4:
-                counts['unknown'] = ','.join(unk)
-            table.add_line(counts)
-        return table
-
-
-class TrajNotebookMixin(object):
-    @utils.kwargs_from(mdtwidgets.trajectory.TrajectoryViewer)
-    def draw3d(self, **kwargs):
-        """TrajectoryViewer: create a trajectory visualization
-
-        Args:
-            **kwargs (dict): kwargs for :class:`moldesign.widgets.trajectory.TrajectoryViewer`
-        """
-        self._viz = mdtwidgets.trajectory.TrajectoryViewer(self, **kwargs)
-        return self._viz
-    draw = draw3d  # synonym for backwards compatibility
-
-    def draw_orbitals(self, align=True):
-        """ Visualize trajectory with molecular orbitals
-
-        Args:
-            align (bool): Align orbital phases (i.e., multiplying by -1 as needed) to prevent sign
-               flips between frames
-
-        Returns:
-            TrajectoryOrbViewer: create a trajectory visualization
-        """
-        for frame in self:
-            if 'wfn' not in frame:
-                raise ValueError("Can't draw orbitals - orbital information missing in at least "
-                                 "one frame. It must be calculated with a QM method.")
-
-        if align: self.align_orbital_phases()
-        self._viz = mdtwidgets.trajectory.TrajectoryOrbViewer(self)
-        return self._viz
-
-    def plot(self, x, y, **kwargs):
-        """ Create a matplotlib plot of property x against property y
-
-        Args:
-            x,y (str): names of the properties
-            **kwargs (dict): kwargs for :meth:`matplotlib.pylab.plot`
-
-        Returns:
-            List[matplotlib.lines.Lines2D]: the lines that were plotted
-
-        """
-        from matplotlib import pylab
-        xl = yl = None
-        if type(x) is str:
-            strx = x
-            x = getattr(self, x)
-            xl = '%s / %s' % (strx, x.units)
-        if type(y) is str:
-            stry = y
-            y = getattr(self, y)
-            yl = '%s / %s' % (stry, y.units)
-        plt = pylab.plot(x, y, **kwargs)
-        pylab.xlabel(xl); pylab.ylabel(yl); pylab.grid()
-        return plt
